Change the MediaFileUplaod object to be resumable.
Was running into a write error when trying to upload a test mp4 file. YouTube Data API documentation suggests setting the resumable flag to True in the MediaFileUpload object sent in the request media body. This allows for large media files like videos to be uploaded in chunks rather than all at once, which resolves the previously stated error.