Create simple html page for Google OAuth
Our uploader currently uses the command line for oauth authorization.
This authentication is needed to upload videos to youtube channels. We
want to be able to do authorization on the client side then pass the
access token through the controller to the uploader. Use the gapi
javascript client to allow for this client side authorization.