Move api key and client id to gitignored keys file
Having the api key and client id in the javascript is unavoidable.
Opening dev tools in the browser, we can see the keys.js file and its
values. We need to have the api key and client id in the javascript in
order to run oauth. We mitigate this by setting restrictions on the api
key and client id from the Google API dashboard. Regenerating the keys
and moving them to a gitignore file is an added precaution.

In order to properly load the keys js file, we place the html and
javascript in appropriate folders according to Flask documentation.